refactor CurvedBottom
